Try it out for yourself
If you know a way to make the computer to tell you if you are close to the number within a certain range to display a new message. For example the number generated is 5 and you guess 4, I want to have the display message to be different than the ordinary "higher" or "lower". Instead saying something like "almost there"